product class exception handle api

+++ src/main/java/com/pabaro/ims/service/impl/ProductServiceImpl.java
@@ -2,6 +2,7 @@ package com.pabaro.ims.service.impl;
 
 import com.pabaro.ims.dto.Product;
 import com.pabaro.ims.entity.ProductEntity;
+import com.pabaro.ims.exception.ProductNotFoundException;
 import com.pabaro.ims.repository.ProductJpaRepository;
 import com.pabaro.ims.repository.ProductRepository;
 import com.pabaro.ims.service.ProductService;
@@ -42,17 +43,29 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public Product retrieveById(Long id) {
         Optional<ProductEntity> productEntity = repository.findById(id);
+
+        if (repository.findById(id).isEmpty())
+            throw new ProductNotFoundException(String.format("%d No product found with this ID",id));
+
         return new ModelMapper().map(productEntity, Product.class);
     }
 
     @Override
     public void update(Product product) {
         ProductEntity productEntity = new ModelMapper().map(product, ProductEntity.class);
+
+        if (repository.findById(product.getId()).isEmpty())
+            throw new ProductNotFoundException(String.format("%s This Product Not Found",product));
+
         repository.update(productEntity);
     }
 
     @Override
     public void deleteById(Long id) {
+
+        if (repository.findById(id).isEmpty())
+            throw new ProductNotFoundException(String.format("%d No product found with this ID",id));
+
         repository.deleteById(id);
     }
 }
